Adds a quick option menu and a function for menu validation.
The menu is hooked on the shorten button on the toolbar, which is designed for
fast switching between shortening services and check if to ignore native shortlinks.

This commit adds the menu to the toolbarItem, and implements a function to set correct
checkedState for every menuItem when the menu is about to display (which fires a `menu` event.)

`SafariExtensionMenu` and `SafariExtensionMenuItem` are introduced in Safari 5.1,
so this quick option menu is only available in Safari 5.1 and later.

TODO:
  - Implement the commands of menuItems
  - Add locale capability