Shortly 2.0 beta 2
This build contains:
  - New feature: a keyboard hotkey for shortening.
  - New feature: a quick option menu on the toolbar shorten button (Safari 5.1+ only; details below)
  - Improved Italiano locale thanks to Raffaele Viggiani
  - Improved Simplified Chinese locale thanks to @catjia1990
  - Updating jsOAuth library to ver 1.3
  - Other bug fixes and enhancements

About Quick option menu
-----------------------
The menu is hooked on the shorten button on the toolbar, which is designed for
fast switching between shortening services and check if to ignore native shortlinks.

`SafariExtensionMenu` and `SafariExtensionMenuItem` are introduced in Safari 5.1,
so this quick option menu is only available in Safari 5.1 and later.